Move repeated functions to utils file
