+++ server/src/actions/start-commercial.js
@@ -1,5 +1,8 @@
 const { ApiClient } = require("@twurple/api");
 const { StaticAuthProvider } = require("@twurple/auth");
+const { getTwitchChannel,
+    getTwitchAPIClient
+} = require("../action-utils");
 module.exports = {
     key: "start-commercial",
     auth: ["client"],
@@ -16,20 +19,8 @@ module.exports = {
      */
     // eslint-disable-next-line no-empty-pattern
     async handler(success, error, { commercialDuration }, { client }, { get, auth }) {
-        const broadcast = await get(client?.broadcast?.[0]);
-        if (!broadcast) return error("No broadcast associated");
-        if (!broadcast.channel) return error("No channel associated with broadcast");
-
-        const channel = await auth.getChannel(broadcast?.channel?.[0]);
-        if (!channel.twitch_refresh_token) return error("No twitch auth token associated with channel");
-        if (!channel.channel_id || !channel.name || !channel.twitch_scopes) return error("Invalid channel data");
-        let scopes = channel.twitch_scopes.split(" ");
-        if (!["channel:edit:commercial"].every(scope => scopes.includes(scope))) return error("Token doesn't have the required scopes");
-
-        const accessToken = await auth.getTwitchAccessToken(channel);
-
-        const authProvider = new StaticAuthProvider(process.env.TWITCH_CLIENT_ID, accessToken);
-        const api = new ApiClient({authProvider});
+        const { channel } = getTwitchChannel(client, ["channel:edit:commercial"], { get, auth }, { success, error });
+        const api = getTwitchAPIClient(channel, auth);
 
         try {
             await api.channels.startChannelCommercial(channel.channel_id, commercialDuration);
